refactor: move core function to new file
